feat: adds basic validation with model load
Adds option to validate when loading a model from
a io.Reader. A noop validation is added for testing or when
skipping validation is desired. A schema validator is added for validation
against the OSCAL JSON schemas.

The implemented validation logic is meant to be a post-decoding semantic
validation of the models where a validation interface is added. The goal is to
allow customized validation including combined validations. Adding the schema validation after
unmarshaling to a struct has some performance overhead to the requirement to marshal
back to JSON for validation against the schema. This may be reevaluated, but proposing this
trade off be accepted for additional validation flexibility.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>