feat!: add control source to assessment plan reviewed controls (#76)
* feat!: add control source to assessment plan backmatter

To ensure an Assessment Plan created from a set
of Component Definitions can still be traced back to the original
control source, the information from the control implementation is
added as a backmatter resource.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: typofix in framework_test.go

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: add link to backmatter under reviewed controls

Adding a link under reviewed controls makes it clear to
the consumer why the OSCAL Profile is being added and referenced.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates example output in IncludeControls

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>